requirements.txt
streamlit
pandas
opencv-python-headless
pytesseract
Pillow
openpyxl